requirements.txt
hydra-core==1.3.2
ipython==8.11.0
librosa==0.9.1
matplotlib==3.7.1
numba==0.56.4
numpy==1.23.5
omegaconf==2.3.0
praat-parselmouth==0.4.3
protobuf==3.20.0
pytorch-lightning==2.0.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.4
streamlit==1.22.0
tensorboard==2.12.0
torch==2.0.0
torchaudio==2.0.1
torchcrepe==0.0.18
tqdm==4.65.0
transformers==4.27.2
typing_extensions==4.8.0
google-cloud-storage==2.8.0
fleep==1.0.1
autoroot
typer==0.9.0
evaluate==0.4.0
jiwer==3.0.2
Unidecode==1.3.6
inflect==7.0.0
wavinfo==2.2.1